requirements.txt
Django==3.1.2
line-bot-sdk==1.17.0
gunicorn==20.0.4
pandas==1.1.5
jieba==0.42.1
lxml==4.6.3
ntplib==0.4.0